Rename uploader file
